backend/requirement.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
GeoAlchemy2==0.18.4
geopandas==1.1.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
numpy==2.4.4
packaging==26.0
pandas==3.0.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
